Embed version via bundle
