+++ packages/bundle/.babelrc
@@ -1,6 +1,7 @@
 {
   "plugins": [
-    "@babel/proposal-object-rest-spread"
+    "@babel/proposal-object-rest-spread",
+    "babel-plugin-version-transform"
   ],
   "presets": [
     "@babel/preset-typescript",
