+++ packages/bundle/package.json
@@ -62,6 +62,7 @@
     "@types/node": "^10.9.4",
     "@types/react": "^16.4.13",
     "babel-core": "7.0.0-bridge.0",
+    "babel-plugin-version-transform": "^1.0.0",
     "concurrently": "^4.0.1",
     "typescript": "^3.0.3",
     "webpack": "^4.17.2",
