+++ packages/component/package.json
@@ -48,6 +48,7 @@
     "@types/react": "^16.4.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^23.0.1",
+    "babel-plugin-version-transform": "^1.0.0",
     "botframework-directlinejs": "0.9.18-master.3cece81",
     "concurrently": "^3.5.1",
     "jest": "^22.4.4",
