+++ packages/core/.babelrc
@@ -1,7 +1,8 @@
 {
   "plugins": [
     "@babel/proposal-object-rest-spread",
-    "@babel/plugin-transform-runtime"
+    "@babel/plugin-transform-runtime",
+    "babel-plugin-version-transform"
   ],
   "presets": [
     "@babel/preset-typescript",
