+++ packages/core/package.json
@@ -46,6 +46,7 @@
     "@types/node": "^10.3.3",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^23.0.1",
+    "babel-plugin-version-transform": "^1.0.0",
     "botframework-directlinejs": "0.9.18-master.3cece81",
     "concurrently": "^3.5.1",
     "jest": "^22.4.4",
